+++ src/meta-content.c
@@ -46,12 +46,12 @@
 
 extern struct _Rule_Struct *rulestruct;
 
-bool Meta_Content(int rule_position, const char *syslog_message)
+bool Meta_Content(uint_fast32_t rule_position, const char *syslog_message)
 {
 
-    int z=0;
-    int meta_alter_num=0;
-    int match=0;
+    uint_fast8_t z=0;
+    uint_fast32_t meta_alter_num=0;
+    uint_fast8_t match=0;
 
     char meta_alter_content[MAX_SYSLOGMSG] = { 0 };
 
@@ -145,7 +145,7 @@ bool Meta_Content(int rule_position, const char *syslog_message)
 /* Meta_Content_Search does the actual "searching" (or content!) of the data */
 /*****************************************************************************/
 
-bool Meta_Content_Search(char *syslog_msg, int rule_position, int meta_content_count)
+bool Meta_Content_Search(char *syslog_msg, uint_fast32_t rule_position, uint_fast8_t meta_content_count)
 {
 
     int z = meta_content_count;
